fixed graphical bugs, implemented 8.4

+++ src/hitrecord/mod.rs
@@ -6,7 +6,7 @@ use crate::vec3::Vec3;
 pub struct HitRecord {
     pub p: Point,
     pub normal: Vec3,
-    pub t: f32,
+    pub t: f64,
     pub front_face: bool,
 }
 
@@ -19,7 +19,7 @@ impl HitRecord {
             front_face: false,
         }
     }
-    fn from(p: Point, normal: Vec3, t: f32, front_face: bool) -> Self {
+    fn from(p: Point, normal: Vec3, t: f64, front_face: bool) -> Self {
         HitRecord {
             p,
             normal,
@@ -28,7 +28,7 @@ impl HitRecord {
         }
     }
 
-    #[inline(always)]
+    #[inline]
     pub fn set_face_normal(&mut self, r: &Ray, outward_normal: Vec3) {
         self.front_face = Vec3::dot(&r.direction, &outward_normal) < 0.0;
         self.normal = if self.front_face {
